pyproject.toml
[build-system]
requires = ["setuptools>=72.1.0", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "SnakeAltPromoter"
version = "1.0.5"
description = "A Snakemake pipeline for alternative promoter analysis"
readme = "README.md"
license = { file = "LICENSE.md" }
authors = [
    { name = "Yidan Sun", email = "syidan@wustl.edu" },
]
maintainers = [
    { name = "Jared Lalmansingh", email = "jared.lalmansingh@wustl.edu" },
]
classifiers=[
    "Programming Language :: Python :: 3.11",
    "Operating System :: POSIX",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "snakemake>=8.28.0",
    "streamlit",
    "pyarrow",
    "pandas",
    "pyyaml",
]
requires-python = ">=3.11"


[project.scripts]
Genomesetup = "snakealtpromoter.workflows.Genomesetup:main"
Snakealtpromoter = "snakealtpromoter.workflows.Snakealtpromoter:main"
sap = "snakealtpromoter.cli:main"
sap-ui = "snakealtpromoter.ui.launch:main"


[tool.setuptools]
include-package-data = true
packages = [
    "snakealtpromoter",
    "snakealtpromoter.data",
    "snakealtpromoter.data.samplesheet",
    "snakealtpromoter.docs",
    "snakealtpromoter.rules",
    "snakealtpromoter.rules.envs",
    "snakealtpromoter.scripts",
    "snakealtpromoter.scripts.downstream_plots",
    "snakealtpromoter.ui",
    "snakealtpromoter.workflows",
]


[tool.setuptools.package-data]
snakealtpromoter = [
    "data/*",
    "data/samplesheet/*",
    "docs/*",
    "rules/*.Snakefile",
    "rules/envs/*.yaml",
    "ui/*.py",
    "scripts/*",
]


[project.urls]
homepage = "https://github.com/YidanSunResearchLab/SnakeAltPromoter"
repository = "https://github.com/YidanSunResearchLab/SnakeAltPromoter"
issues = "https://github.com/YidanSunResearchLab/SnakeAltPromoter/issues"


[project.optional-dependencies]
test = [
    "pytest",
]
